feat: #5852 updates billing tab
